Don't allow the 'self' argument in __init__ to be annotated.
Report [invalid-annotation] if we find an annotation on 'self' in __init__ in
a Python file. Such an annotation doesn't really make sense, since __init__ is
used specifically for initializing the class it's defined by. This change is
safe for us to make - there isn't currently any code in [] that would run
afoul of the new error: []

Once we've reported the error, we delete the annotation. This way, an __init__
call with fake args always succeeds, so we can remove the problematic
try/except (context: []#comment4).

--
PiperOrigin-RevId: 175325965
MOS_MIGRATED_REVID=175325965