Handle circular tuple types in _value_to_parameter_types().
As odd as it is, when you stick a tuple inside another tuple, then create an
enumerate object followed by a closure of the tuple, pytype ends up with a
tuple representation that contains itself. With this change, we'll set all of
the tuple's element types to Any when we detect this case.

--
PiperOrigin-RevId: 175844954
MOS_MIGRATED_REVID=175844954