Add a --strict_none flag to pytype.
Adds a --strict_none flag that turns on strict attribute checking for None.
When ths flag is on, pytype reports attribute errors on None even if other
possible types are present. We filter out false positives by not reporting
errors for which the path back to the None assignment passes through the
inference supernode. This filtering strategy fails in one obvious case:

def f():
  pass
def g():
  f().upper()  # Clearly should be an error.

for which I've added a skipped test. I think we can fix this problem by
adding a node for byte_RETURN_VALUE, connecting the fake inference supernode
to it, and blocking off the node before it, but I still need to flesh that
idea out.

--
PiperOrigin-RevId: 176006696
MOS_MIGRATED_REVID=176006696