Copy a variable's origins when reusing a cached function return value.
I'm pretty sure I was mistaken in thinking that the old origins needed to be
thrown away. Since we're reanalyzing the call in the same environment, those
origins should still be valid.

This change is needed for --strict_none, which depends on origins being
correct, but I can't formulate a test until some other issues with
--strict_none are worked out.

--
PiperOrigin-RevId: 176573030
MOS_MIGRATED_REVID=176573030