Copy cached return values more carefully in InterpreterFunction calls.
When we re-use a cached return value, we need to be more thoughtful about how
we copy the origins. This bug has been in pytype for a while - --strict_none
just exposes it. For the longest time we simply threw the origins away, and
recently I switched to naively copying all of them, neither of which works.
Instead, we should throw away the origins that are within the cached call but
keep the ones from outside it.

--
PiperOrigin-RevId: 177604133
MOS_MIGRATED_REVID=177604133