Allow to create a load_pytd.Loader without an options data structure.
We use load_pytd in a lot of places, including several that don't have a natural
way to create an option dictionary. It's cleaner to explicitly pass the
parameters to the loader. In particular, this will enable us to get rid of
GetBuiltinsAndTyping and instead retrieve builtins/typing from the loader.

--
PiperOrigin-RevId: 177656233
MOS_MIGRATED_REVID=177656233