Use --strict_none to filter retrieved attributes instead of errors.
Previously, --strict_none had the problem that when pytype infers types for
something like:
  x = None
  def f():
    return x
we lose the information that the return type of f() comes from a module level
constant, so we get false positives again if we put
  def f() -> None
in the pyi file.

We should filter the return type of the function and output
  def f() -> Any
. But if we try to do the filtering in, say, byte_RETURN_VALUE or
InterpreterFunction.call(), we run into all sorts of issues with generators,
containers, etc. Instead, I've change load_from() and _retrieve_attr() to
filter the attribute values they return.

The nice thing about this change is that we no longer need to call
has_strict_none_origins() during error logging, since we've already filtered
out the false positives.

--
PiperOrigin-RevId: 178031468
MOS_MIGRATED_REVID=178031468