Update cfg.py and cfg_utils.py so that we can switch between cfg.cc/py.
I've found it useful for debugging to switch back to cfg.py in order to be
able to add print statements or run pdb. Adds Program.is_reachable to cfg.py
and changes set.union(...) to itertools.chain(...) in cfg_utils.py to be
compatible with both sets and frozensets.

--
PiperOrigin-RevId: 178391563
MOS_MIGRATED_REVID=178391563