Fix a bug in how solver.cc checks for goal conflicts.
When checking for conflicts in removed goals, we need to also include the
current goals. The reason is that if we have two goals, one of which has an
origin that conflicts with the other, and that origin is removed by
RemoveFinishedGoals(), then the only way to find the conflict is to call
GoalsConflict() with both the removed origin and the conflicting other goal.

--
PiperOrigin-RevId: 178998550
MOS_MIGRATED_REVID=178998550