Keep content origins when creating a list.
When an abstract.List is created, we need to keep the origins for the contents
instead of copying the values to the root node.

A side effect of this change is that we can get rid of both the node and
keep_origins arguments to initialize_type_parameter.

This change also exposed a mistake in a test in test_typing.py, which is fixed.

--
PiperOrigin-RevId: 181108104
MOS_MIGRATED_REVID=181108104