Partially support typing.NoReturn.
Adding NoReturn allows pytype to type-check the following:

x = ...  # type: None or X
if not x:
  function_that_raises_exception()
x.attribute_of_X

Otherwise, pytype does not realize that x cannot be None at the last line.
This CL
* Adds NoReturn to typing.pytd.
* Infers a type of NoReturn if a function always raises an exception.
* Sets state.exception to True if a function calls another that cannot return.
We get type-checking against NoReturn for free.

The only part of NoReturn that this CL doesn't implement is verifying that
NoReturn doesn't appear in places other than as the return type of a function.

--
PiperOrigin-RevId: 182277755
MOS_MIGRATED_REVID=182277755