Make it an error to not return from a function with an annotated return type.
Moves some of the logic in byte_RETURN_VALUE into a reusable function that
checks the return value and, if applicable, replaces it with the return
annotation. This function is then used by byte_RAISE_VARARGS and the various
function call opcodes (through call_function_from_stack).

Deletes a test for
  class Foo(object):
    def __new__(cls) -> "Foo":
      assert cls is not Foo, "not instantiable"
I added this test because I saw this pattern in pytree.py in our test data and
thought we might want to support it, but on further thought, anyone who wants
a non-instantiable base class should just use abc.ABCMeta.

--
PiperOrigin-RevId: 182453578
MOS_MIGRATED_REVID=182453578