Don't report bad-return-type if NoReturn is one among many options.
I previously put the return type checking for NoReturn immediately after an
exception was raised, so that we'd have the right frame and traceback. An
unintentional side effect was that we'd report a bad return type even if we
also found an acceptable return value. Instead, save the stack information and
do the checking after we know that NoReturn is the only option.

--
PiperOrigin-RevId: 182560984
MOS_MIGRATED_REVID=182560984