Remove --strict_none check from _filter_none_and_paste_bindings.
This fixes a bug that arises due to an interaction between late initialization
and NoReturn:

class Foo(object):
  def __init__(self):
    self.foo = None
  def get_foo(self) -> int:
    if self.foo:
      return 42
    else:
      raise ValueError()

Without this change, pytype reports [bad-return-type] for get_foo because it
thinks that 'return 42' is unreachable.

The change to main_test is necessary because a (real) attribute error in
pytree.py was hidden by a similar late initialization pattern.

--
PiperOrigin-RevId: 182801442
MOS_MIGRATED_REVID=182801442