Add __len__ to various builtin classes.
Sized.__len__ is an abstract method, so we need to add a concrete __len__ to
instantiable classes; otherwise, it's possible to use multiple inheritance to
trick pytype into thinking subclasses of some builtins aren't instantiable.

Also adds __getitem__ to Dict for the same reason, and removes Iterable from
the parents of buffer because buffer implements __getitem__, not __iter__, and
that confuses pytype.

--
PiperOrigin-RevId: 183867061
MOS_MIGRATED_REVID=183867061