Remove non-determinism from pytypes saving file functionality.
From:
"The mtime argument is an optional numeric timestamp to be written to the stream when compressing. All gzip compressed streams are required to contain a timestamp. If omitted or None, the current time is used. This module ignores the timestamp when decompressing; however, some programs, such as gunzip, make use of it. The format of the timestamp is the same as that of the return value of time.time() and of the st_mtime attribute of the object returned by os.stat()."

This CL passes the mtime parameter to an old but stable value.
As we use gzip to unzip the files and the value is therefore never read this is not a problem.

I used three runs of /google/data/ro/projects/devtools/blaze/check_determinism.par -- devtools/python/blaze/pytype:precompiled_builtins_2 to check that this fix works. The actual test used some more patches to make check_determinsm behave like --define PYTYPE=DEV would have been passed.

Also unknown commit was patched in.

--
PiperOrigin-RevId: 184278615
MOS_MIGRATED_REVID=184278615