Fix a bug in our error message printing.
Found while trying to change how abstract.List initializes its type parameter.
For
  lst = [str]
  lst[0] = [int]
we printed '[str or int]', which deceptively looks correct, but we don't
actually do position-aware __setitem__ on lists - what's happening is that
we're adding an int binding to the pyval entry (which is also a bug - to be
fixed in an upcoming change). The best pytype can do in this case is
'instance of List[Type[int or str]]'.

--
PiperOrigin-RevId: 184707359
MOS_MIGRATED_REVID=184707359