Don't confuse module aliases with classes with the same name.
LookupExternalTypes is used on both asts parsed from pyi files
(which can contain module aliases) and the ast inferred by pytype
(which cannot). In the latter case, if the Python file contains:
  from foo import bar
  class foo: ...
pytype puts 'foo' in the module map, so - even though module foo
isn't in the namespace - pytype confuses module foo and class foo.
Avoid this problem by introducing a module alias only when we're
actually inside an alias and the name is fully-qualified (which
means the ast was parsed).

--
PiperOrigin-RevId: 184899701
MOS_MIGRATED_REVID=184899701