Handle Python 3 bytestrings properly in loadmarshal.py.
Create a str subclass, BytesPy3, to mark Python 3 bytestrings, and teach
convert.py to convert BytesPy3 to an instance of __builtin__.bytes.

Also changes loadmarshal_test.py to test both Python 2 and 3 when a version
isn't specified for loadmarshal.loads(), and to check types when comparing
strings, since assertEqual() treats "foo" and u"foo" as equal. This change
catches a bug in the test - test_load_interned and test_load_unknown tested
the same input but declared different expected outputs ('test' vs. u'test').

Changes 'bytes("foo", "utf-8")' to 'b"foo"' in all of the Python 3 tests that
used the former only because of this bug.

--
PiperOrigin-RevId: 185560192
MOS_MIGRATED_REVID=185560192