Don't check 'self' when matching protocol methods.
We failed to match PyTD classes against parameterized protocols because we
would, e.g., match
  def __iter__(self: Foo): ...
against
  def __iter__(self: Iterable[T]): ...
and pytype would try to enforce contravariance of the 'self' argument.

Fixing this bug flushed out another that caused us to think that List
satisfies the Mapping protocol, so I added a check in _match_against_protocol
to prevent that.

--
PiperOrigin-RevId: 185920569
MOS_MIGRATED_REVID=185920569