Fix a corner case in which Instance.__init__ could cause infinite recursion.
For a class definition like
  class Foo(ContextManager[Foo], Iterable[bytes]): ...
where ContextManager and Iterable share the same type parameter, we give the
parameter a value of Any. Using merge_type_parameter() meant that we evaluated
the parameter's value while overwriting it with Any, so Foo tried to
instantiate ContextManager[Foo], crashing pytype.

Instead, we'll overwrite the dict value directly, explicitly using
dict.__setitem__ so that the type parameters dict doesn't complain about our
replacing the variable (which is usually a bad idea).

--
PiperOrigin-RevId: 186573240
MOS_MIGRATED_REVID=186573240