Change the result of type(self) in __init__ to Any.
This is an imperfect solution - it only works for __init__, and it
means that if-splitting will no longer work for things like:

def __init__(self):
  if type(self) is Foo:
    self.x = ...
  else:
    self.x = ...

However, extending the solution to methods other than __init__ would
make the code more complicated, since we would no longer be able to
to take advantage of the existing callself_stack attribute. I don't
see a way to keep the if-splitting.

--
PiperOrigin-RevId: 187350630
MOS_MIGRATED_REVID=187350630