Make --strict_none the default again.
I'll submit this as soon as we reach a state in which pytype looks ready for
release (which is hopefully the current state, but I'm doing a
build_all_pytype_targets to check).

I've left none_attr in errors.py because I realized that we need to delete
occurrences of it in [] before removing it from our errors, or users
will be bombarded with error messages about an invalid directive.

Also makes a couple of minor changes suggested by gpylint.

--
PiperOrigin-RevId: 187546313
MOS_MIGRATED_REVID=187546313