Special-case List against Mapping in protocol matching.
Detecting that a list isn't a mapping is especially tricky because it has all
of the right methods; as far as I can tell, we have to compare

__getitem__(self, int) -> T (list)
__getitem__(self, K) -> V (mapping)

and

__iter__(self) -> Iterator[T] (list)
__iter__(self) -> Iterator[K] (mapping)

and notice that T is mapped to both V and K, which is illegal. I'd like to
insert this workaround for now because the previous code was preventing
functools.partial from matching Callable, which generated at least 39 false
positives when I tried to prepare a release.

--
PiperOrigin-RevId: 187587006
MOS_MIGRATED_REVID=187587006