Rollback of commit a6538a0d630c9a2f87a758f53a8fa4032d08c425.
*** Reason for rollback ***

I jumped the gun on submitting this change, as I discovered two bugs that
needed to be fixed before doing a release. Those have been fixed, now, and
I did a build_all_pytype_targets with this rollback patched in to confirm
that everything looks reasonable.

Note that this isn't quite a pure rollback - the original change contained
two lint changes, and I've kept those to keep pylint happy.

*** Original change description ***

Make --strict_none the default again.

I'll submit this as soon as we reach a state in which pytype looks ready for
release (which is hopefully the current state, but I'm doing a
build_all_pytype_targets to check).

I've left none_attr in errors.py because I realized that we need to delete
occurrences of it in [] before removing it from our errors, or users
will be bombarded with error messages about an invalid directive.

Also makes a couple of minor changes suggested by gpylint.

***

--
PiperOrigin-RevId: 187612510
MOS_MIGRATED_REVID=187612510