Be more judicious in our use of AdjustSelf(force=True).
Previously, we were indiscriminately using force=True in
various places, which is problematic now that we we want
to add annotations like
  def __enter__(self: T, ...) -> T
since AdjustSelf would remove the custom annotation.

This changed uncovered a couple of incorrect tests, which
I've fixed.

Also deletes an unnecessary call to CanonicalOrdering()
that I noticed in test_base.

--
PiperOrigin-RevId: 188348139
MOS_MIGRATED_REVID=188348139