Replace class types with bounded type params when necessary.
Changes the CreateTypeParametersForSignatures visitor to replace
occurrences of a class X with TypeVar(bound=X) in all signatures
that match (self: X[, ...) -> X.

The downside of doing this replacement in a visitor is that we
sometimes replace things we shouldn't - this is an issue for
classes that define a __copy__ method, since the two common ways
of copying an instance of X are type(self)(...) and X(...), and
the latter we'll no longer analyze correctly.

The upsides are that
(1) it's a simple, clean change, and
(2) we start inferring the right signatures for both
    ContextManager.__enter__ and namedtuple._replace. (For the
    latter, what happened was that I had to tweak collections_overlay
    to unbreak a test, and then realized that my tweak actually
    improved the signature of _replace.)
So I think this change is a net improvement.

--
PiperOrigin-RevId: 188351572
MOS_MIGRATED_REVID=188351572