Improve parser_ext error handling
Calling PyObject_CallObject with an error set triggers an assert under python3.

Save and restore the error instead of bailing early.
If we exit before calling the function, we don't set the line number.

But if we restore the error unconditionally, and there wasn't an error before we called the object *and* the call raises an exception, we'd wind up clobbering the exception with the empty error we stored initially, causing a "SystemError: error return without exception set"

Tested:

--
PiperOrigin-RevId: 188501817
MOS_MIGRATED_REVID=188501817