pytype: Support `unicode` annotations for Python 2/3 compatibility
When porting code to support both Python 2 and Python 3, it is common to
add various __future__ declarations, including:

  from __future__ import unicode_literals

This makes Python 2 default to all string literals being `unicode`
instead of `str`, because this better approximates the Python 3
environment.

Unfortunately, when "string" annotations are used for delayed evaluation
of pytype annotations, we get an error from code like this:

  class Foo(object):
    @classmethod
    def MakeFoo() -> 'Foo': ...

Because pytype doesn't think that annotations can be unicode, it reports:

  Invalid type annotation "u'Foo'" for return [invalid-annotation]

As a fix, pytype needs to understand the `unicode` type in Python 2, and
be able to recognize it when it is used as an annotation literal.

This requires fixing the `namedtuple` code to actually handle unicode field names, instead of silently ignoring them as unparsable constants.

--
PiperOrigin-RevId: 189593330
MOS_MIGRATED_REVID=189593330