Make pytd/serialize_ast.py 2-and-3 compatible.
This module accesses the message attribute on a builtin exception several
times, so I created a utils.message() function for this access.

I *think* it's better to use utils.message() everywhere rather than
also having a mixin to do the same thing, but I don't feel strongly about this,
so if one of you would prefer to keep the mixin, I'll add it back.

--
PiperOrigin-RevId: 191958081
MOS_MIGRATED_REVID=191958081