Convert test_calls and test_checker to 2and3.
Tweaks loadmarshal to return native strings from load_string() except when
loading the code and lnotab fields of a CodeType object. That load_string()
is used for both bytes and native strings is a problem I also ran into when
converting test_slots, although that time I tried to fix the problem later on,
by accepting bytes for slot names, which turned out to be the wrong solution.

--
PiperOrigin-RevId: 192802600
MOS_MIGRATED_REVID=192802600