Make Text an alias for Union[str, unicode] and don't auto-promote unicode.
PEP 484 calls Text 'a simple alias for [...] unicode' but also says that
'[f]or parameters typed as Text, arguments of type str as well as unicode
should be acceptable'. We've been treating the unicode annotation as
equivalent to Union[str, unicode], which satisfies the spec but prevents users
from ever specifying a type that is just unicode. Instead, make unicode
just unicode and use Text for Union[str, unicode].

--
PiperOrigin-RevId: 193590017
MOS_MIGRATED_REVID=193590017