Trim down the opensource scrubber's replacements dict.
We were using the scrubber both to make necessary code replacements and to
sanitize comments. Removes everything that was in the latter category and
disables scrub_sensitive_comments so that leak_test will yell at us if we
write something that isn't safe for export. Also fixes a leaky comment that was
previously being scrubbed.

--
PiperOrigin-RevId: 195026037
MOS_MIGRATED_REVID=195026037