Dedup the mutations we apply after a function call.
A dictionary of tuples generates a lot of views during argument
matching, which means that a dict.setitem call leads to the
same mutation being applied repeatedly to a type parameter,
creating a complicated CFG that pytype later chokes on.

With this change, we'll build a set of mutations to apply instead
of a list.

In [], I noted that there's some problem with our C++
solver that causes it to be much slower than cfg.py in rare cases.
That was the case in [] as well, and I haven't figured out
why it happens - this fix simply avoids creating the problematic CFG.

--
PiperOrigin-RevId: 195695384
MOS_MIGRATED_REVID=195695384