Update pytype-all to reflect changes in importlab.
Also includes the following changes:

Add opensource filegroups to the new blaze packages in the tests directory.

If a directory is a subpackage, then filegroup "srcs" in the parent package will not
include files glob-ed from that directory. We need to add a filegroup target in the
subpackage and list this target in the "srcs" field of the parent package filegroup.

--
PiperOrigin-RevId: 195741649
MOS_MIGRATED_REVID=195741649