Update pytype_runner's behavior.
Updates pytype_runner to report errors for a file it analyzes iff the file
is in the filenames for which the user asked for type-checking. Combines the
'projects' and 'deps' options into a single 'pythonpath' option, now that
'projects' isn't needed to determine when to report errors.

Changes pyi_dir to be pytype_output/ itself rather than pytype_output/pyi/,
since pytype_output/ no longer contains a log file.

Moves some logic out of run() and run_pytype() into yield_sorted_modules()
for testing.

For sanity's sake, changes pytype-all to do all of its path expansion
upfront, so that pytype_runner doesn't need to worry about expanding paths,
and tests know that they always have to do the expansion themselves.

While updating the README, I noticed that the section on filling in a config
file made it seem like the package directory itself should be in the
pythonpath, so I tweaked the language and example to be clearer.

--
PiperOrigin-RevId: 197109460
MOS_MIGRATED_REVID=197109460