Make the unit tests executable via the command line unittest command.
The unit tests cannot be run currently with a command like:

python -m unittest ...

This change enables one to execute the unit tests with the above command.

Unit tests cannot be run with the above command currently because it does
not run the main test module as a Python script. It imports the main test
module and does its own magic over the tests defined in the module. This
does not work for us as we want test_base.main() to be called.

--
PiperOrigin-RevId: 197175001
MOS_MIGRATED_REVID=197175001