Make sure that pytype checks ancestry for abc.ABCMeta-ness.
Core issue appeared to be that pytype would only check to see if the type was abc.ABCMeta. This meant that in cases where you created a child of abc.ABCMeta, pytype would complain. The fix is then pretty striaghtforward: look up the mro for each metaclass to see if it is a child of abc.ABCMeta.

--
PiperOrigin-RevId: 198055934
MOS_MIGRATED_REVID=198055934