Bug fix: don't expand default paths relative to a config file.
Fixes a bug where if a config file was found, pythonpath and output_dir
would be expanded relative to the file's directory regardless of whether
they were specified in the file. Instead, we pass the expansion functions
to ConfigSection's keymap so that only those paths that come from
ConfigSection are expanded.

This change also lays the groundwork for doing path expansion on command
line args.

--
PiperOrigin-RevId: 198193828
MOS_MIGRATED_REVID=198193828