Create convenience methods vm.PY2 and vm.PY3.
A few weeks ago, I promised to do this in a "followup CL". A little
late, but better than never?

Adds convenience methods for quickly checking the target Python version.
This change only makes it easier to check for major versions, not
minor ones, and only for the VM's python version, but even so, it
covers the majority of our version-checking.

Also fixes a few lint errors and gets rid of a pointless tuple() call.

--
PiperOrigin-RevId: 198418654
MOS_MIGRATED_REVID=198418654