Exposes some pytype-single options to pytype-all.
Splits the previously comment-based categories in make_parser() into separate
functions, including an add_basic_options() that we reuse in pytype-all's
parser. (As a side note, you were right that pulling --check and --output out
of basic options into their own category is useful.)

--disable, --no-report-errors, and --protocols can now be set from pytype-all's
command line. While --quick and --analyze-annotated should also be available,
they're not as easy to get and not as important as --disable, so I punted on
them for now. The next step is to integrate the new variables into Config so
that they can be set from a file.

--
PiperOrigin-RevId: 198801843
MOS_MIGRATED_REVID=198801843