Add py_test CMake targets.
This change does not add a corresponding py_test target for all of the existing
py2and3_test targets. The CMake py_test rule should be extended (for example, add
a MAIN option) to add test targets for all of them. It will be done in a future CL.

Some numbers:
To run all of the tests without ninja (which means sequentially) on my machine, it
took around 35 minutes. With ninja running the tests in parallel,  it takes only a
minute.

--
PiperOrigin-RevId: 198930725
MOS_MIGRATED_REVID=198930725