Allow pytype-single args to be specified in a config file.
* Tweaks config.Postprocessor to make in-place processing easier.
* Adds a SimpleNamespace class to datatypes, since multiple tests need the
  ability to quickly create namespaces.
* Changes the way Config objects work so that all the variable processing
  happens on separate objects, then the final attributes are copied to Config.
  This makes it easier to do complex processing without, e.g., defaults and
  file options interfering with each other.
* Creates a parse_args.Parser class to do parsing, postprocessing, and initial
  config file creation.

The steps for constructing a Config object are now as follows:
(1) Parser.parse_args() creates a new args object with all of the command-line
    values for config.ITEMS and the pytype-single args. config.ITEMS'
    converters are automatically called as part of parsing, and parse_args()
    calls postprocess() for the pytype-single args.
(2) A new Config is created and populated with defaults by
    Parser.defaults_to_config(). Again, config.ITEMS' converters come for free,
    and the Parser method calls postprocess() for the pytype-single args.
(3) File options are read into a FileConfig object. config.ITEMS' converters
    are called as part of read_from_file(), and main.py calls postprocess() for
    the pytype-single args. The FileConfig is merged into the Config object.
(4) The args object from (1) is merged into the Config object.

The reason I put the bulk of the new logic in parse_args.py rather than
config.py is that config is imported by parse_args, and I didn't want to create
a logical loop by passing the parser back to config.

--
PiperOrigin-RevId: 198950273
MOS_MIGRATED_REVID=198950273