Adjust the scrubber directive removing "from __future__ import google_type_annotations".
The directive is now a regex directive which removes all spaces before the "from"
keyword. Also, it keeps the ending newline character so that line numbers are
preserved after the scrubber transformation. Preserving line numbers is required to
ensure that we have errors on the same line numbers in [] and outside.

--
PiperOrigin-RevId: 199213284
MOS_MIGRATED_REVID=199213284