Optimization: skip argument views that we've already seen.
When we create variable->binding views for a function's arguments
to analyze a call, we include every variable that may possibly be
used, so we often generate a lot of redundant views.

This change tracks which items in a view are actually used, so if
we see another view that contains the same items, we can skip it.

This optimization doesn't do much in the average case (no change
in Kessel Run times, as far as I can tell), but it helps a lot in
some pathological ones. For example, we can now check the file
on the release blacklist in about 30s.

--
PiperOrigin-RevId: 200083493
MOS_MIGRATED_REVID=200083493