Move QualifyRelativeNames into the parser.
The original motivation for this change was to mitigate what looked like
a performance regression caused by my previous change to
QualifyRelativeNames. I'm no longer sure whether this regression was real
or random noise in the Kessel Run times, but this CL makes our code nicer,
and removing the overhead of a visitor won't slow us down, at least.

This change:
* Replaces QualifyRelativeNames with a _qualify_name method in the parser.
  Note that the builtins rename was already being (accidentally) done by the
  visitor; we're just being more explicit about it now.
* Gets rid of the now-unused TypeDeclUnit.is_package.
* Fixes some lint errors.

Tested with:
* blaze test third_party/py/pytype/...
* blaze build third_party/py/pytype:type_checked_lib --define PYTYPE=DEV
* blaze test --define PYTYPE=DEV devtools/python/proto_to_pyi:all\
  devtools/python/clif_to_pyi:all devtools/python/blaze/pytype:all\
  devtools/python/blaze/pytype/scripts:all
* Unpacking blaze-bin/third_party/py/pytype/opensource/public-tree.tar into
  a local copy of the pytype repository, pip installing from the repo, then
  running typeshed/tests/pytype_test.

--
PiperOrigin-RevId: 200734430
MOS_MIGRATED_REVID=200734430