Drop Python references to binding data when the Program instance goes away.
In case of the cfg extension module, binding data are Python objects owned by C++
Binding instances. The binding data can also be potentially shared between multiple
Binding instances. With this CL, management of ownership and cleanup of such shared
binding data (which in case of Python objects is to drop a reference to them) is done
using std::shared_ptrs with custom deleters. The Python references to the bound
Python objects are dropped when they are no longer needed.

PiperOrigin-RevId: 206093558