Switch AtomicAbstractValue.cls from a variable to a value.
The motivation for this change was to reduce the number of
cfg.Variables per pytype run. Since solving typegraph queries
is our main performance bottleneck, I figured that shrinking
the graph might help a bit.

The only reason for cls to be a variable was to support setting __class__,
which is an extremely rare operation (done in .07% of Python files in Google).
Making it a value not only reduces the number of variables created by about
20% as measured on pytree.py but also considerably simplifies some code.

This change does the bare minimum needed to make the switch from
variable to value; I've left TODOs for further simplifications
that this switch enables.

I ran a build_all_pytype_targets and everything looks good:
target,failure,[]_link
//platforms/testing/infra/meetup/pyclient:demo,Findings,[]

PiperOrigin-RevId: 206822962