Clean up two more TODOs related to attribute.py
Resolves the remaining TODOs from my AtomicAbstractValue.cls change. Because
the different parts of attribute handling are so tightly interconnected, I
had to put all the behavior modifications in the same change; a follow-on one
will do a bunch of small refactorings/aesthetic improvements.

This change reduces the number of variables created while analyzing pytree.py
by a further 20% (of the original count).

Along the way, I tried to document as much as possible, since I came across a
bunch of hidden assumptions (most of which were probably introduced by me...).

* Changes property_get to accept one callself argument plus an is_class flag,
  rather than callself and callcls. The two were effectively mutually
  exclusive, which was unclear.
* Changes BoundFunction and friends to no longer require a callcls
  parameter to their constructor.
* Fixes a CallTracer.bind_method call that was passing in a binding instead
  of a variable for the instance parameter.
* Deletes get_attribute_generic (now unused).
* Gets rid of the valcls argument in attribute.py.
* Replaces the logic of creating a value getter and passing it to
  _get_value_or_class_attribute with the simpler approach of having the latter
  method determine based on its inputs what the getter should be.

Blacklists three files for which this change causes new findings. All of them
appear to be issues that pytype already knew how to catch, so this change has
probably fixed some subtle bug by simplifying our code.

#Squeegee

PiperOrigin-RevId: 207934024