Fix a use-after-free in CFGNode names.
PyObject_Str creates a new Python str object, but PyString_AsString returns a
pointer to the Python object's internal buffer. That buffer is deallocated when
the Python object is deallocated, leading to a use-after-free when the CFGNode
is created. This was found by running ASAN on cfg_test.py, but was only found
on the Python 3 version of the test.

Also fixes a possible DECREF of a null pointer found by clang_tidy.

PiperOrigin-RevId: 208125733