Add cpython submodule pinned to the current 2.7 HEAD.
Extended the build script to build the patched Python2.7 interpreter if
the option "--py27" is specified. A standalone script which can build
the patched interpreter has also been added. This script will be used by
a CMake target to build the patched interpreter as part of the
run_tests.py flow.