Make utils_test work under Python-3.7.
A test method in this test module assumes that an expression
like id((1,)) == id((1,)) evaluates to False. However, Python-3.7
seems to be doing a more aggressive constant literal caching, which
makes that expression to evaluate to True. So, this change is
essentially making the test method evaluate id of a sufficiently
complex expression to avoid caching.

PiperOrigin-RevId: 211901694