Support setting pytype-all's --exclude argument in the config file.
Adds --exclude to config.ITEMS. Adds a compat.recursive_glob() function to
support this change. Also does two tiny cleanups: adds a newline after
`[pytype]` in the sample config, for readability, and deletes an extra
whitespace in the help message for --verbosity.

I mostly developed this feature in the GitHub project and also ran it there on
a few simple examples, but with this much code, I felt more comfortable
linting and testing it internally.

PiperOrigin-RevId: 213486175