Show a full sample.cfg.
Wraps up my work on the config file bug by adding the pytype-single arguments
to sample.cfg and making it a fully functional sample config file - i.e.,
someone could generate it and immediately use it for their project.

Also makes sure the documentation on the config file is up-to-date. Since I was
touching the README, I moved the "not an official Google product" disclaimer to
the end - Adam had pointed out that having it at the beginning may be scaring
people off.

PiperOrigin-RevId: 213909618