Filter closure variables at load time.
Calls cfg.Variable.Filter() in load_closure_cell so that variables are
filtered before we put them in a closure, since when we analyze a closure on
its own, we're no longer at the right node to do filtering.

Also fixes a LOAD_GLOBAL bug (None being loaded as Any) exposed by this change.

PiperOrigin-RevId: 216580334