Get rid of uses of LazyDict's methods.
Gets rid of uses of lazy_eq and add_lazy_item, with the goal of
switching from using a LazyDict for type_parameters to using a regular
dict, lazily loaded. I ran build_kessel_run over this change, and it
has no performance impact, as far as I can tell.

* In instantiate(), rather than checking for parameter names in
  type_parameters, checks in the template, which is more accurate and
  avoids using type_parameters before it is fully loaded.
* Adds __eq__ and __hash__ methods to TypeParameterInstance and Tuple
  to get rid of lazy_eq calls. Improves ParameterizedClass.__hash__.
* Delays loading of type_parameters to get rid of add_lazy_item calls.
* Renames Class._get_template to _compute_template to match existing
  compute_* methods and to avoid having two _get_template functions.

PiperOrigin-RevId: 219196752