Remove custom email notification settings from .travis.yml.
Based on my reading of https://docs.travis-ci.com/user/notifications/,
I believe that without the custom settings, the email frequency will be
the same, but the emails will be sent to the PR author (if the author
is a member of the pytype repo). This seems good enough to me and will
stop our spamming of the external mailing list.

Fixes #185.