Cleanup: make PyTDFunction.get_constructor_args a classmethod.
This is the first in a string of CLs aimed at breaking up abstract.py. The
overall plan is:
* Make as many things as possible stand-alone functions.
* Move everything except AtomicAbstractValue and its subclasses out
  of abstract.py.
* Put MixinMeta and its subclasses in a new mixin.py, function-related things
  in function.py, and everything else in a new abstract_utils.py.

Part of the first step is turning every static method into either a
stand-alone function (so it can be moved) or a class method (to prove that it
should continue to be in abstract.py).

This change turns PyTDFunction.get_constructor_args into a classmethod. It
fixes some lint errors along the way.

PiperOrigin-RevId: 221357473