Cleanup (part 5): Fix the import relationship between pytd_utils and parser.
Previously, parser depended on pytd_utils in BUILD, but pytd_utils imported
parser. Removes the problematic import by moving the definition of JoinTypes to
pytd_utils and that of canonical_pyi to parser.

This change will allow pytd_for_parser to be depended on by upcoming targets
that don't pull in the parser.

PiperOrigin-RevId: 221541060