Fix a bug where pytype would fail to resolve an unused import.
This is a problem when a pyi files does `from X import Y as Y` with
the intent to re-export Y. pytype would see Alias(Y, X.Y) and assume
that X was a module and Y a module attribute. Instead, we need to
detect when X is a package and attempt to load a pyi for X.Y.

When this is LGTM'd, I'll update __version__ and CHANGELOG with
the current date before submitting, so that I can do a PyPI release
and unblacklist the affected third_party typeshed files.

Makes progress toward https://github.com/google/pytype/issues/171.

PiperOrigin-RevId: 247335339